refactor(cli): give lib/terminal a single public entry
